La inn ting som manglet + flyttet yup valideringer inn i egen fil
