+++ apps/frontend/src/components/DpProcess/DpProcessModal.tsx
@@ -7,7 +7,7 @@ import BoolField from '../Process/common/BoolField'
 import CustomizedModalBlock from '../common/CustomizedModalBlock'
 import FieldProduct from '../common/FieldProduct'
 import { Error, ModalLabel } from '../common/ModalSchema'
-import { dpProcessSchema } from '../common/schema'
+import { dpProcessSchema } from '../common/schemaValidation'
 import FieldDescription from './common/FieldDescription'
 import FieldDpProcessAffiliation from './common/FieldDpProcessAffiliation'
 import FieldDpProcessDataProcessingAgreements from './common/FieldDpProcessDataProcessingAgreements'
