+++ apps/frontend/src/components/Process/Accordion/CardLegalBasis.tsx
@@ -18,7 +18,7 @@ import {
 } from '../../../service/Codelist'
 import { LegalBasisView } from '../../common/LegalBasis'
 import { paddingZero } from '../../common/Style'
-import { legalBasisSchema } from '../../common/schema'
+import { legalBasisSchema } from '../../common/schemaValidation'
 import { customizeNationalLawPlaceholder } from './PlaceholderCustomizer'
 
 const Error = (props: { fieldName: string }) => (
@@ -102,7 +102,7 @@ const CardLegalBasis = ({
   return (
     <Formik
       onSubmit={(values) => submit(values)}
-      validationSchema={legalBasisSchema()}
+      validationSchema={legalBasisSchema(codelistUtils)}
       initialValues={initialValues}
       render={(form: FormikProps<ILegalBasisFormValues>) => (
         <Card>
