+++ apps/frontend/src/components/Process/Accordion/ModalPolicy.tsx
@@ -15,15 +15,13 @@ import {
 import { Fragment, useEffect, useState } from 'react'
 import { getInformationTypesShort } from '../../../api/GetAllApi'
 import { ELegalBasesUse, IInformationTypeShort, IPolicyFormValues } from '../../../constants'
-import { CodelistService, EListName } from '../../../service/Codelist'
+import { EListName, ICodelistProps } from '../../../service/Codelist'
 import { disableEnter } from '../../../util/helper-functions'
 import CustomizedStatefulTooltip from '../../common/CustomizedStatefulTooltip'
 import { Error, ModalLabel } from '../../common/ModalSchema'
-import { policySchema } from '../../common/schema'
+import { policySchema } from '../../common/schemaValidation'
 import FieldLegalBasis from '../common/FieldLegalBasis'
 import { TDocs } from './TablePolicy'
-
-const [codelistUtils] = CodelistService()
 
 const renderTagList = (list: string[], arrayHelpers: FieldArrayRenderProps) => (
   <Fragment>
@@ -136,6 +134,7 @@ type TModalPolicyProps = {
   submit: (values: IPolicyFormValues) => void
   onClose: () => void
   addBatch?: () => void
+  codelistUtils: ICodelistProps
 }
 
 const ModalPolicy = ({
@@ -147,6 +146,7 @@ const ModalPolicy = ({
   docs,
   title,
   addBatch,
+  codelistUtils,
 }: TModalPolicyProps) => (
   <Modal
     onClose={onClose}
@@ -159,7 +159,7 @@ const ModalPolicy = ({
     <div className="w-[750px] px-8">
       <Formik
         initialValues={initialValues}
-        validationSchema={policySchema()}
+        validationSchema={policySchema(codelistUtils)}
         onSubmit={(values: IPolicyFormValues) => {
           submit(values)
           onClose()
