+++ apps/frontend/src/components/Processor/ProcessorModal.tsx
@@ -12,7 +12,7 @@ import BoolField from '../Process/common/BoolField'
 import PanelTitle from '../Process/common/PanelTitle'
 import CustomizedModalBlock from '../common/CustomizedModalBlock'
 import { Error, ModalLabel } from '../common/ModalSchema'
-import { dataProcessorSchema } from '../common/schema'
+import { dataProcessorSchema } from '../common/schemaValidation'
 import FieldContractOwner from './components/FieldContractOwner'
 import FieldCountries from './components/FieldCountries'
 import FieldNote from './components/FieldNote'
