+++ apps/frontend/src/test/schema/process.test.ts
@@ -1,12 +1,9 @@
-import { processSchema } from '../../components/common/schema'
+import { processSchema } from '../../components/common/schemaValidation'
 import { EProcessStatus, ILegalBasisFormValues, IProcessFormValues } from '../../constants'
-import { EListName, NATIONAL_LAW_GDPR_ARTICLES } from '../../service/Codelist'
+import { CodelistService, EListName, NATIONAL_LAW_GDPR_ARTICLES } from '../../service/Codelist'
 import { addCode } from '../config/codelist'
 import '../config/schemaValidator'
 
-addCode(EListName.PURPOSE, 'PURPOSE')
-
-const schema = processSchema()
 export const createProcess = (): IProcessFormValues => ({
   name: 'name',
   purposes: ['PURPOSE'],
@@ -34,41 +31,49 @@ export const createProcess = (): IProcessFormValues => ({
   disclosures: [],
 })
 
-test('Process ok', () => {
-  expect(createProcess()).toBeSchema(schema)
-})
+it('Process', () => {
+  addCode(EListName.PURPOSE, 'PURPOSE')
 
-test('Process purposes required', () => {
-  const process = { ...createProcess(), purposes: [] }
-  expect(process).toBeSchemaErrorAt(schema, 'purposes')
-})
+  const [codelistUtils] = CodelistService()
 
-test('Process status needs to be a type ProcessStatus', () => {
-  const process = { ...createProcess(), status: '' }
-  expect(process).toBeSchemaErrorAt(schema, 'status')
-})
+  const schema = processSchema(codelistUtils)
 
-test('Process disclosures required', () => {
-  const process = { ...createProcess(), disclosures: '' }
-  expect(process).toBeSchemaErrorAt(schema, 'disclosures')
-})
+  test('Process ok', () => {
+    expect(createProcess()).toBeSchema(schema)
+  })
 
-test('Process legalBasis', () => {
-  let process = {
-    ...createProcess(),
-    legalBases: [{ gdpr: NATIONAL_LAW_GDPR_ARTICLES[0], description: '', nationalLaw: '' }],
-  }
-  expect(process).toBeSchemaErrorAt(schema, 'legalBases[0].description')
-  process = {
-    ...createProcess(),
-    legalBases: [{ gdpr: NATIONAL_LAW_GDPR_ARTICLES[0], description: 'desc', nationalLaw: '' }],
-  }
-  expect(process).toBeSchemaErrorAt(schema, 'legalBases[0].nationalLaw')
-  process = {
-    ...createProcess(),
-    legalBases: [
-      { gdpr: NATIONAL_LAW_GDPR_ARTICLES[0], description: 'desc', nationalLaw: 'LAW_1' },
-    ],
-  }
-  expect(process).toBeSchema(schema)
+  test('Process purposes required', () => {
+    const process = { ...createProcess(), purposes: [] }
+    expect(process).toBeSchemaErrorAt(schema, 'purposes')
+  })
+
+  test('Process status needs to be a type ProcessStatus', () => {
+    const process = { ...createProcess(), status: '' }
+    expect(process).toBeSchemaErrorAt(schema, 'status')
+  })
+
+  test('Process disclosures required', () => {
+    const process = { ...createProcess(), disclosures: '' }
+    expect(process).toBeSchemaErrorAt(schema, 'disclosures')
+  })
+
+  test('Process legalBasis', () => {
+    let process = {
+      ...createProcess(),
+      legalBases: [{ gdpr: NATIONAL_LAW_GDPR_ARTICLES[0], description: '', nationalLaw: '' }],
+    }
+    expect(process).toBeSchemaErrorAt(schema, 'legalBases[0].description')
+    process = {
+      ...createProcess(),
+      legalBases: [{ gdpr: NATIONAL_LAW_GDPR_ARTICLES[0], description: 'desc', nationalLaw: '' }],
+    }
+    expect(process).toBeSchemaErrorAt(schema, 'legalBases[0].nationalLaw')
+    process = {
+      ...createProcess(),
+      legalBases: [
+        { gdpr: NATIONAL_LAW_GDPR_ARTICLES[0], description: 'desc', nationalLaw: 'LAW_1' },
+      ],
+    }
+    expect(process).toBeSchema(schema)
+  })
 })
